requirements.txt
# Core application framework
streamlit

# Data handling
pandas

# Environment variables
python-dotenv

# OpenAI for summary generation
openai>=1.3.0

# Tool integration SDKs
arcadepy
vapi-python

# Advanced web search and content extraction
httpx
trafilatura

# Geographic mapping (MSA) and NLP utilities
spacy==3.7.2
rapidfuzz
scikit-learn
python-dateutil
beautifulsoup4
lxml